// src/types/generated/PuzzlePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PuzzleSpecificData } from "./PuzzleSpecificData";

export type PuzzlePayload = {
  id: string;
  title: string;
  grid: (string | number | null)[][];
  specificData: PuzzleSpecificData;
  gridFont?: string;
  titleFont?: string;
  themeAccents?: boolean;
  cellBorders?: boolean;
  ideTheme?: boolean;
  letterTracking?: number;
  wordBankColumns?: number;
  selectorStyle?: string;
  solutionStyle?: string;
};